FinalProject/backend/requirements.txt
ag2[openai]
flask
flask-cors
python-dotenv
